Fix loss=inf and NCCL timeout: use float32 focal loss + gloo without DDP
- CenterNetFocalLoss: cast to float32 before sigmoid to prevent 1-1e-6
  rounding to 1.0 in AMP float16, which caused log(0) = -inf
- Switch from NCCL+DDP to gloo without DDP wrapper (backbone frozen,
  each GPU trains independently on its own shard subset)
- Per-rank seeding for data diversity across GPUs
- Lightweight dist.barrier() between epochs for sync

+++ ultrassl/lesion_localizer.py
@@ -142,8 +142,10 @@ class CenterNetFocalLoss(nn.Module):
         Returns:
             Scalar loss.
         """
-        pred = torch.sigmoid(pred_logits)
+        # Force float32 — in AMP float16, 1-1e-6 rounds to 1.0 → log(0) = -inf
+        pred = torch.sigmoid(pred_logits.float())
         pred = pred.clamp(1e-6, 1 - 1e-6)
+        target = target.float()
 
         pos_mask = target.eq(1).float()
         neg_mask = target.lt(1).float()
